First draft of the continuation-based GLL combinators finished.
These are closely based on Spiewak's Scala combinators but with major
adaptations to make them work in Python and to add support for
semantic actions, combinators that call themselves, and regexes and
binary data.  There's still a lot of work to be done in adapting them
to run fast on Python and to add major missing features.  I also fixed
some bugs in related files.