Went back to a binary sequence combinator to fix bugs due to Spiewak's changes to GLL.
In the standard version of the GLL algorithm, the GSS stores the
position within a nonterminal, what they call the grammar slot.
Spiewak's version doesn't because it's forcing all nonterminals to
be length 2 and using separate continuations for processing the
first grammar slot and the second grammar slot.  Allowing the use of
arbitrary-length sequences will require modifying the GSS handling.